WIP: Move render blocking JavaScript to bottom of page
By default JavaScript blocks DOM construction and thus delays
the time to first render

Moving this to the bottom of the page delays the JavaScript load but
will improve the render time.

OpenLibrary makes extensive use of inline scripts inside the page that
make use of $.ready. However, if jQuery is loaded at the bottom of the
page, then there will be no jQuery at time of execution.

Instead of running the jQuery function we push our function callbacks to
a ol.q which has is defined at the very top of the page. This provides an
array of callbacks to run on execution.

This global queue now gets initialised inside our Open library init
function.