Move render blocking JavaScript to bottom of page
By default JavaScript blocks DOM construction and thus delays
the time to first render

Moving this to the bottom of the page delays the JavaScript load but
will improve the render time.

To make this transition easier, and to deal with all the inline scripts
inside the page, we make use of a global queue which gets cleared as
soon as the footer and jQuery has loaded. We use ES5 forEach function.
Can make use of a for loop if a larger browser support matrix is needed