Add a lightweight OpenLibrary bootstrap code
To ensure we can load JS and CSS off the critical path while not
breaking the existing approach to inline scripting in OpenLibrary
we introduce a small lightweight library that should be loaded before
initialising the site.

We load this as part of the vendor build, with the intention that
all code inside build/all.js will be loaded asynchronously.