Selected shelf on "want to read" is deterministic
Removing a book from a shelf should use a different option
value from the other shelves - otherwise it's not clear to the browser
which option value to render as selected.

Right now we're relying on the first result to be the selected one
but this is what the selected property is for!

The template is DRYed up to be more readable and use the selected property
rather than rely on element order.

The API is modified to accept and remove an entry from its current
shelf when the value -1 is passed.

Fixes: #1123